SSL: enable hash_size values for sha224, sha384 and sha512
The original comment is not correct. Although we do not default to any
of those hashes, with TLS 1.2 we negotiate the hash to be used and
dafault to the strongest hash shared by server and client. So we can
very well end up with any of those hashes since we announce support for
them (see ssl_handshake:default_hash_signs/0)